requirements.txt
# IoT Predictive Maintenance Edge - Dependencies
# Core ML and data processing
numpy>=1.24.0,<2.0.0
pandas>=2.0.0,<3.0.0
scikit-learn>=1.3.0,<2.0.0
scipy>=1.11.0,<2.0.0

# Configuration and validation
pyyaml>=6.0,<7.0
pydantic>=2.0.0,<3.0.0

# Testing
pytest>=7.4.0,<9.0.0
pytest-cov>=4.1.0,<6.0.0
